astro-engine/pyproject.toml
[project]
name = "astro-engine"
version = "0.1.0"
description = "Planetary phase-curve correlation engine for market analysis"
requires-python = ">=3.11"
dependencies = [
    "pyswisseph>=2.10",
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "plotly>=5.18",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[tool.pytest.ini_options]
testpaths = ["tests"]
